% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{signatureTypes}
\alias{signatureTypes}
\title{A list of the allowed signatures type}
\format{
An object of class \code{list} of length 3.
}
\usage{
signatureTypes
}
\description{
This list contains all the allowed signature types.
Use alexandrov2 for Alexandrov V2 signatures, 
alandrov32 for Alexandrov V3.2 ones and shiraishi for Shiraishi signatures.
}
\keyword{datasets}
