% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTumor.R
\name{tumorHeatmap}
\alias{tumorHeatmap}
\title{Computes the exposure vectors from an MPF file and plots 
an heatmap clustering the different tumor genomes by their 
similarities in the calculated exposure different tumarl genomes by their}
\usage{
tumorHeatmap(
  mpfFilePath,
  signatures,
  signaturesType = signatureTypes$alexandrov32,
  numBases = 5,
  trDir = TRUE,
  enforceUniqueTrDir = TRUE,
  refGenome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  transcriptAnno = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  verbose = FALSE
)
}
\arguments{
\item{mpfFilePath}{(Mandatory) The path of the MPF file from which to load the tumor
genomes}

\item{signatures}{(Mandatory) Either a path to the signature file or a vector of 
path to signatures file (in the case of Shiraishi signatures). Also, it can be
dircetly an object containing the signatures. The object can be A list of 
vectors, data frames or matrices. Vectors are used for Alexandrov signatures, 
data frames or matrices for Shiraishi signatures.}

\item{trDir}{(Mandatory) Specifies whether the transcription direction is 
taken into account in the signature model. If so, only mutations within
genomic regions with a defined transcription direction can be 
considered. Default: TRUE}

\item{enforceUniqueTrDir}{(Optional) Used only if trDir is TRUE. If
enforceUniqueTrDir is TRUE (default), then mutations which map to a region
with multiple overlapping genes with opposing transcription directions will
be excluded from the analysis. If FALSE, the transcript direction 
encountered first in the transcript database (see transcriptAnno) is
assigned to the mutation.}

\item{refGenome}{refGenome    (Mandatory) The reference genome (BSgenome)
needed to extract sequence patterns. Default: BSgenome object for hg19.}

\item{transcriptAnno}{(Optional) Transcript annotation (TxDb object) used to
determine the transcription direction. This is required only if trDir is 
TRUE. Default: TxDb object for hg19.}

\item{verbose}{(Optional) Print information about reading and processing the
mutation data. Default: TRUE}

\item{signaturesTypes}{(Optional) A string representing the signature type
a list of signature types is available as the object `signatureTypes`}
}
\value{
The exposure vector is returned and a heatmap of their similarity is
plotted
}
\description{
Computes the exposure vectors from an MPF file and plots 
an heatmap clustering the different tumor genomes by their 
similarities in the calculated exposure different tumarl genomes by their
}
\details{
Many of the parameters of this function are also described in the function
`readGenomesFromMPF{decompTumor2Sig}` @seealso [readGenomesFromMPF()] 
 @seealso [decomposeTumorGenomes()] as that function and package are at the
base of this function and many parameteres are passed directly to those
functions. A more detailed explanation of the parameters and the function can 
be found there.
}
\seealso{
[signatureTypes]. Use alexandrov2 for Alexandrov V2 signatures, 
alandrov32 for Alexandrov V3.2 ones and shiraishi for Shiraishi signatures.
}
