% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{signatureTypes}
\alias{signatureTypes}
\title{A list of the available signatures}
\format{
An object of class \code{list} of length 3.
}
\usage{
signatureTypes
}
\description{
A list of the available signatures
}
\keyword{datasets}
